+++ src/main/java/com/beholder/watch/dtos/usecases/HttpResponseDetails.java
@@ -3,7 +3,7 @@ package com.beholder.watch.dtos.usecases;
 import com.beholder.watch.model.watchable.WatchableStatus;
 
 public interface HttpResponseDetails {
-  public int getResponseTime();
+  public long getResponseTime();
   public int getResponseStatus();
   public WatchableStatus getWatchableStatus();
   public String getErrorMessage();
